+++ src/status.rs
@@ -1,9 +1,16 @@
 //! Module for long lasting status effects.
 
-use crate::battle::BattleRules;
-use crate::event::EventId;
+use crate::battle::{Battle, BattleRules};
+use crate::character::verify_is_character;
+use crate::entity::EntityId;
+use crate::error::WeaselResult;
+use crate::event::{Event, EventId, EventKind, EventProcessor, EventQueue, EventTrigger};
 use crate::fight::FightRules;
 use crate::util::Id;
+#[cfg(feature = "serialization")]
+use serde::{Deserialize, Serialize};
+use std::any::Any;
+use std::fmt::{Debug, Formatter, Result};
 
 /// A long lasting effect altering an entity's condition.
 ///
@@ -68,5 +75,155 @@ impl<R: BattleRules> std::ops::Deref for LinkedStatus<R> {
 impl<R: BattleRules> std::ops::DerefMut for LinkedStatus<R> {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.status
+    }
+}
+
+/// An event to inflict a status effect on a character.
+///
+/// # Examples
+/// ```
+/// use weasel::battle::{Battle, BattleRules};
+/// use weasel::event::{EventTrigger, EventKind};
+/// use weasel::status::InflictStatus;
+/// use weasel::{Server, battle_rules, rules::empty::*};
+///
+/// battle_rules! {}
+///
+/// let battle = Battle::builder(CustomRules::new()).build();
+/// let mut server = Server::builder(battle).build();
+///
+/// todo
+/// ```
+#[cfg_attr(feature = "serialization", derive(Serialize, Deserialize))]
+pub struct InflictStatus<R: BattleRules> {
+    #[cfg_attr(
+        feature = "serialization",
+        serde(bound(
+            serialize = "EntityId<R>: Serialize",
+            deserialize = "EntityId<R>: Deserialize<'de>"
+        ))
+    )]
+    entity_id: EntityId<R>,
+
+    #[cfg_attr(
+        feature = "serialization",
+        serde(bound(
+            serialize = "StatusId<R>: Serialize",
+            deserialize = "StatusId<R>: Deserialize<'de>"
+        ))
+    )]
+    status_id: StatusId<R>,
+
+    #[cfg_attr(
+        feature = "serialization",
+        serde(bound(
+            serialize = "Option<Potency<R>>: Serialize",
+            deserialize = "Option<Potency<R>>: Deserialize<'de>"
+        ))
+    )]
+    potency: Option<Potency<R>>,
+}
+
+impl<R: BattleRules> InflictStatus<R> {
+    /// Returns a trigger for this event.
+    pub fn trigger<'a, P: EventProcessor<R>>(
+        processor: &'a mut P,
+        entity_id: EntityId<R>,
+        status_id: StatusId<R>,
+    ) -> InflictStatusTrigger<'a, R, P> {
+        InflictStatusTrigger {
+            processor,
+            entity_id,
+            status_id,
+            potency: None,
+        }
+    }
+
+    /// Returns the id of the entity target of this status.
+    pub fn entity_id(&self) -> &EntityId<R> {
+        &self.entity_id
+    }
+
+    /// Returns the id of the status to be inflicted.
+    pub fn status_id(&self) -> &StatusId<R> {
+        &self.status_id
+    }
+
+    /// Returns the status' potency.
+    pub fn potency(&self) -> &Option<Potency<R>> {
+        &self.potency
+    }
+}
+
+impl<R: BattleRules> Debug for InflictStatus<R> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> Result {
+        write!(
+            f,
+            "InflictStatus {{ entity_id: {:?}, status_id: {:?}, potency: {:?} }}",
+            self.entity_id, self.status_id, self.potency
+        )
+    }
+}
+
+impl<R: BattleRules> Clone for InflictStatus<R> {
+    fn clone(&self) -> Self {
+        InflictStatus {
+            entity_id: self.entity_id.clone(),
+            status_id: self.status_id.clone(),
+            potency: self.potency.clone(),
+        }
+    }
+}
+
+impl<R: BattleRules + 'static> Event<R> for InflictStatus<R> {
+    fn verify(&self, battle: &Battle<R>) -> WeaselResult<(), R> {
+        verify_is_character(battle.entities(), &self.entity_id)
+    }
+
+    fn apply(&self, battle: &mut Battle<R>, event_queue: &mut Option<EventQueue<R>>) {
+        // TODO
+    }
+
+    fn kind(&self) -> EventKind {
+        EventKind::InflictStatus
+    }
+
+    fn box_clone(&self) -> Box<dyn Event<R>> {
+        Box::new(self.clone())
+    }
+
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+}
+
+/// Trigger to build and fire an `InflictStatus` event.
+pub struct InflictStatusTrigger<'a, R, P>
+where
+    R: BattleRules,
+    P: EventProcessor<R>,
+{
+    processor: &'a mut P,
+    entity_id: EntityId<R>,
+    status_id: StatusId<R>,
+    potency: Option<Potency<R>>,
+}
+
+impl<'a, R, P> EventTrigger<'a, R, P> for InflictStatusTrigger<'a, R, P>
+where
+    R: BattleRules + 'static,
+    P: EventProcessor<R>,
+{
+    fn processor(&'a mut self) -> &'a mut P {
+        self.processor
+    }
+
+    /// Returns an `InflictStatus` event.
+    fn event(&self) -> Box<dyn Event<R>> {
+        Box::new(InflictStatus {
+            entity_id: self.entity_id.clone(),
+            status_id: self.status_id.clone(),
+            potency: self.potency.clone(),
+        })
     }
 }